refactor(imu): address CodeRabbit round 5 findings
Apply 19 CodeRabbit round 5 review findings to the BNO055/BMP280 IMU
integration branch. All changes preserve full firmware build and test
suite (49/49 tests pass).

Key changes by finding:

Fix 1 (telemetry.proto): Add Hamilton-convention quaternion mapping
comment to quat_w field documenting ROS2 geometry_msgs/Quaternion layout.

Fix 2 (shared_data.h): Reorder imu_state_t fields to place uint32_t
timestamp_ms before 8-bit fields (int8_t/uint8_t/bool), eliminating
compiler padding and reducing struct size.

Fixes 3+4 (rx_bno055.h): Extract bno055_calib_shift_t, bno055_calib_mask_t,
and bno055_scale_t enums from rx_bno055_regs.h to rx_bno055.h. Remove
bno055_calib_t from regs.h (duplicate enumerators conflict). Rename
bno055_scale_t in regs.h to bno055_regs_scale_t to avoid redefinition.

Fix 5 (test_rx_bmp280.c): Rename helper_ prefix functions to internal_
throughout to match STAR naming conventions.

Fix 6 (main.c): Remove double casts (imu_i2c_addr_t)(uint8_t) and keep
single (imu_i2c_addr_t) cast to maintain typedef usage.

Fix 7 (rx_bmp280_regs.h): Add k_bmp280_reg_calib_end = 0x9F enum value
to bmp280_reg_t. Update comments in rx_bmp280.h and rx_bmp280.c to use
the named constant.

Fix 8 (test_rx_bmp280.c): Add test_bmp280_zero_t enum and use
k_temp_zero_centi_degc in TEST_ASSERT_NOT_EQUAL to eliminate magic zero.

Fix 9 (test_rx_bno055.c): Move test_bno055_read_buf_idx_t and
test_bno055_quat_buf_idx_t enums to the Test Constants section, removing
misplaced declarations from inside a docblock.

Fix 10 (rx_bno055.c): Add full Doxygen block for s_bus_name static
constant including @var, @note, @see, and @since tags.

Fix 11 (rx_bno055.c): Remove redundant (type) casts from array indexing
expressions in internal_read_euler, internal_read_quat, internal_read_lia.
Promote local typedef enums (euler_offsets_t, quat_offsets_t, lia_offsets_t)
to file scope to eliminate unused-local-typedef warnings.

Fix 12 (imu_task.c): Add @var g_bus_manager and @since Version 1.0.0 to
the extern declaration Doxygen block.

Fix 13 (rx_bno055.c): Decompose rx_bno055_init into four static helpers:
internal_init_reset_and_wait, internal_init_configure, internal_init_enter_ndof,
internal_verify_chip_id. Each has full Doxygen and forward declaration.

Fix 14 (test_rx_bno055.c): Change test_quat_expected_t underlying type
from uint16_t to int16_t. Remove redundant (int16_t) casts in assertions.

Fix 15 (rx_bno055.c): Add _Static_assert to internal_assemble_int16_le
verifying shift constant and uint16_t bit width. Add #include <limits.h>
for CHAR_BIT.

Fix 16 (rx_bmp280.c): Add @brief summary comments to all forward
declarations in rx_bmp280.c.

Fixes 17+18+19 (rx_bmp280.c): Add postcondition range checks after
compensation in rx_bmp280_read (Fix 17). Extract internal_parse_u16_le,
internal_parse_s16_le (Fix 18), and internal_assemble_adc20 (Fix 19) as
named inline helpers. Promote local typedef enums to file scope
(bmp280_le16_idx_t, bmp280_adc20_idx_t, bmp280_output_range_t).

Update test_rx_bmp280.c calibration bytes (P1=65410, P2=0) and ADC bytes
so that the mock-constrained adc_P (~9) and the temperature bytes in the
shared mock buffer produce physically valid output within BMP280 operating
range (-40..+85 degC, 300..1100 hPa). Rework test docstrings to explain
the mock RIIC offset-0 constraint and derivation of chosen coefficients.

+++ star-rx72n-firmware/src/main.c
@@ -1718,24 +1718,24 @@ static void internal_register_system_buses(void)
 
   /* Register i2c1 - BNO055 IMU (RIIC1, addr 0x28, SDA=P2.0, SCL=P2.1) */
   err = rx_bus_config_init_i2c(&s_i2c1_imu_config,
-                               "i2c1",                                 /* name: matches rx_bno055.c s_bus_name */
-                               k_riic_channel_1,                        /* channel: RIIC1 */
-                               (uint8_t)(imu_i2c_addr_t)k_i2c_addr_bno055, /* device_addr: BNO055 (COM3/ADR=LOW) */
-                               k_rx_p2_0,                               /* sda_pin: P2.0 = SDA1 */
-                               k_rx_p2_1,                               /* scl_pin: P2.1 = SCL1 */
-                               k_i2c_frequency_400khz);                 /* frequency_hz: 400 kHz fast mode */
+                               "i2c1",                                       /* name: matches rx_bno055.c s_bus_name */
+                               k_riic_channel_1,                             /* channel: RIIC1 */
+                               (imu_i2c_addr_t)k_i2c_addr_bno055,           /* device_addr: BNO055 (COM3/ADR=LOW) */
+                               k_rx_p2_0,                                    /* sda_pin: P2.0 = SDA1 */
+                               k_rx_p2_1,                                    /* scl_pin: P2.1 = SCL1 */
+                               k_i2c_frequency_400khz);                      /* frequency_hz: 400 kHz fast mode */
   RX_ASSERT(err == k_rx_ok, "i2c1 (IMU) config init must succeed");
   err = rx_bus_manager_add_bus(&g_bus_manager, &s_i2c1_imu_config);
   RX_ASSERT(err == k_rx_ok, "i2c1 (IMU) registration must succeed");
 
   /* Register i2c1_baro - BMP280 barometric sensor (RIIC1, addr 0x76, SDA=P2.0, SCL=P2.1) */
   err = rx_bus_config_init_i2c(&s_i2c1_baro_config,
-                               "i2c1_baro",                              /* name: matches rx_bmp280.c s_bus_name */
-                               k_riic_channel_1,                         /* channel: RIIC1 (same as i2c1) */
-                               (uint8_t)(imu_i2c_addr_t)k_i2c_addr_bmp280, /* device_addr: BMP280 (SDO=LOW) */
-                               k_rx_p2_0,                                /* sda_pin: P2.0 = SDA1 */
-                               k_rx_p2_1,                                /* scl_pin: P2.1 = SCL1 */
-                               k_i2c_frequency_400khz);                  /* frequency_hz: 400 kHz fast mode */
+                               "i2c1_baro",                                  /* name: matches rx_bmp280.c s_bus_name */
+                               k_riic_channel_1,                             /* channel: RIIC1 (same as i2c1) */
+                               (imu_i2c_addr_t)k_i2c_addr_bmp280,           /* device_addr: BMP280 (SDO=LOW) */
+                               k_rx_p2_0,                                    /* sda_pin: P2.0 = SDA1 */
+                               k_rx_p2_1,                                    /* scl_pin: P2.1 = SCL1 */
+                               k_i2c_frequency_400khz);                      /* frequency_hz: 400 kHz fast mode */
   RX_ASSERT(err == k_rx_ok, "i2c1_baro (BMP280) config init must succeed");
   err = rx_bus_manager_add_bus(&g_bus_manager, &s_i2c1_baro_config);
   RX_ASSERT(err == k_rx_ok, "i2c1_baro (BMP280) registration must succeed");
